pyproject.toml
[project]
name = "gepa-polyglot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gepa>=0.1.1",
    "grpcio>=1.78.0",
    "grpcio-tools>=1.78.0",
]

[project.scripts]
gepa-rpc = "gepa_rpc.cli:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["gepa_rpc*"]
